Fold the pp-stage guard tests into the loss suite and reuse its helpers
make_parallel_state grows an is_pp_last_stage parameter instead of the test hand-rolling a ParallelState; make_args replaces the local args factory; the file moves next to the other loss tests.
